Add tests cases for the project using PHPUnit
While PHPUnit isn't the perfect tool for testing shell scripts, it does give us the necessary tools to create a temporary filesystem with a pre-determined structure, as well as the ability to overwrite environment variables to scope Asimov to our dummy directory.

In this particular case, we're overloading the tmutil application with a PHP test double, enabling us to track what's called without actually relying on tmutil.